requirements.txt
customtkinter~=5.2.2
matplotlib~=3.10.6
pandas
numpy
python-binance
scikit-learn
torch
mplfinance
